Frankreed/Prevent leave dialog bypass in submit (#174)
* Prevent swipe back (iOS) and back button (Android) from bypassing the leaveDialog

* corrected warning

* update beforeRemove whenever toggleResult changes
