+++ .gitignore
@@ -5,3 +5,6 @@
 /.vs
 /venv
 eas.json
+ios
+android
+webstorm.config.js
