+++ app/segments/drill/[id]/submission/index.js
@@ -1,11 +1,12 @@
-import { useLocalSearchParams } from "expo-router";
-import { useState } from "react";
+import { useLocalSearchParams, useNavigation } from "expo-router";
+import { useEffect, useState } from "react";
 import { DefaultTheme, PaperProvider } from "react-native-paper";
 
 import Loading from "~/components/loading";
 import Input from "./input";
 import Result from "./result";
 
+import DialogComponent from "~/components/dialog";
 import ErrorComponent from "~/components/errorComponent";
 import { useDrillInfo } from "~/hooks/useDrillInfo";
 
@@ -14,6 +15,31 @@ export default function Index() {
 
   const [outputData, setOutputData] = useState([]);
   const [toggleResult, setToggleResult] = useState(false);
+  const [leaveDialogVisible, setLeaveDialogVisible] = useState(false);
+  const hideLeaveDialog = () => setLeaveDialogVisible(false);
+
+  const [exitAction, setExitAction] = useState(null);
+  // Navigation
+  const navigation = useNavigation();
+
+  useEffect(() => {
+    navigation.setOptions({ gestureEnabled: toggleResult });
+  }, [toggleResult]);
+
+  useEffect(() => {
+    return navigation.addListener("beforeRemove", (e) => {
+      if (toggleResult) {
+        return;
+      }
+      // Prevent default behavior of leaving the screen
+      e.preventDefault();
+
+      setExitAction(e.data.action);
+
+      // Prompt the user before leaving the screen
+      setLeaveDialogVisible(true);
+    });
+  }, [toggleResult]);
 
   const {
     data: drillInfo,
@@ -45,5 +71,25 @@ export default function Index() {
     }
   };
 
-  return <PaperProvider theme={DefaultTheme}>{display()}</PaperProvider>;
+  return (
+    <PaperProvider theme={DefaultTheme}>
+      {display()}
+
+      {/* Leave Drill Dialog */}
+      <DialogComponent
+        title={"Alert"}
+        content="All inputs will be lost."
+        visible={leaveDialogVisible}
+        onHide={hideLeaveDialog}
+        buttons={["Cancel", "Leave Drill"]}
+        buttonsFunctions={[
+          hideLeaveDialog,
+          () => {
+            hideLeaveDialog();
+            navigation.dispatch(exitAction);
+          },
+        ]}
+      />
+    </PaperProvider>
+  );
 }
